@@ -1,18 +1,18 @@
 apply plugin: 'com.android.library'
-apply plugin: 'maven-publish'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 36
+    namespace 'com.bytedance.tailor'
 
     defaultConfig {
-        minSdkVersion 14
-        targetSdkVersion 28
+        minSdkVersion 21
+        targetSdkVersion 36
 
         externalNativeBuild {
             cmake {
                 cppFlags "-fvisibility=hidden"
                 abiFilters 'armeabi-v7a', 'arm64-v8a'
-                arguments "-DANDROID_ARM_MODE=arm", "-DANDROID_STL=c++_static"
+                arguments "-DANDROID_ARM_MODE=arm", "-DANDROID_STL=c++_static", "-DANDROID_SUPPORT_FLEXIBLE_PAGE_SIZES=ON"
             }
         }
     }
@@ -24,7 +24,7 @@ android {
         }
     }
 
-    ndkVersion "21.4.7075529"
+    ndkVersion "27.0.12077973"
     externalNativeBuild {
         cmake {
             path 'CMakeLists.txt'
@@ -41,18 +41,5 @@ android {
 
 dependencies {
     implementation fileTree(dir: 'libs', include: ['*.jar'])
-    implementation 'com.android.support:support-annotations:28.0.0'
-}
-
-afterEvaluate {
-    publishing {
-        publications {
-            release(MavenPublication) {
-                from components.release
-                groupId = 'com.github.bytedance'
-                artifactId = 'tailor'
-                version = '1.1.6'
-            }
-        }
-    }
+    implementation 'androidx.annotation:annotation:1.9.1'
 }
